requirements.txt
aiohttp
aiohttp-jinja2
aiosqlite
aiohttp-debugtoolbar
jinja2
APScheduler
click
cogwatch
dataclassy
dateparser
discord.py
discord-ext-menus
discord-py-slash-command
emoji
python-dotenv
humanize
PyYAML
SQLAlchemy
colorlog

#redisent devel package
git+git://github.com/synistree/redisent.git@devel#egg=redisent

#web
flask
flask-Login
flask-SQLAlchemy
flask-Bootstrap
flask-Moment
flask-Pretty
flask-WTF
WTForms-Alchemy
WTForms
Werkzeug

gunicorn
#uwsgi
uvloop
